Add missing action to content blocks controller
The code works without this action because both the route and the view
are specified, and by default Rails renders the view when there's no
action defined.

However, the code is easier to understand if the action is present in
the controller, which is what we do most of the time.