Remove rule MD033 Inline HTML
We aren't following this rule in the `docs/es/features/graphql.md` file
because we need to specify the HTML ID attribute of an element (since by
default markdown generates IDs based on the text, and the text contains
accented letters). There isn't a standard way to do so with regular
markdown and, at least in the past, gitbook hasn't supported syntaxes
that add IDs to elements in markdown.

Ideally we would simply ignore this rule on this file, but I haven't
found a way to do so with markdownlint, so, until we find a better
solution, we're removing this rule.

We've considered keeping the rule and using heading IDs like:

```
 ## Características {#caracteristicas}
```

However, this syntax isn't supported by every application rendering
markdown. For instance, github doesn't support it.