Re-add and apply mdl rule MD040
We were following it about half of the time and we even added it to our
former `.mdlrc` file. However, for some reason, mdl doesn't detect this
rule when specified in the `.mdlrc` file, so we didn't notice we weren't
following it in many cases.

Now that we're using a style file to configure mdl, we can enable this
rule again and apply it, since now mdl correctly includes it in its
report.