Run delayed_job with bundle exec
We've experienced some issues when executing delayed job after upgrading
to version 2.2.0, getting an error on the preproduction server:

```
cannot load such file -- sassc-embedded (LoadError)
```

This line was added to the top of the application.rb file in version
2.2.0. It looks like it isn't properly loaded if we don't use `bundle
exec`.

Since we're changing the delayed job service template, we're making sure
this file is properly updated when deploying.