Upgrade to Rails 7.2
Note we now need to write the `secrets` method, since it's been removed
in Rails 7.2. We're using a simplified version of the `secrets` method
in previous versions of Rails.

Also note we're removing the `secret_key_base` key from the test and
development secrets, since it's ignored in these environments. Quoting
from the Rails documentation [1]:

> In development and test, this is randomly generated and stored in a
> temporary file in <tt>tmp/local_secret.txt</tt>.

Note that, in **development**, existing cookies won't be used after
upgrading. If this is important for you (for example, if you're
constantly switching branches with Consul Democracy 2.4 and Consul
Democracy 2.5 in order to change the upgrade), **on your local
machine**, delete your `secret_key_base` references for the
**development** environment in your `config/secrets.yml`.

Finally, we also need to overwrite the `secret_key_base` method so it
reads the one in `config/secrets.yml`. However, doing so is tricky since
this method contains a few conditionals related to the current
environment and to certain environment variables. So, instead, we're
overwriting the `credentials` method so it returns secrets instead;
since the `secret_key_base` method calls `credentials.secret_key_base`
when running the application on production, overwriting the
`credentials` method will cause the application to get the secret key
base from the `config/secrets.yml` file. Since we don't use credentials
at all, this change doesn't affect the rest of the application.

To avoid confusion, we're changing the Devise initializer so it doesn't
reference the `secrets.secret_key_base` method.

Other minor modifications from the code the `rails app:update` generates
include removing the `APP_NAME` and puma-dev references in `bin/setup`
(they were added in Rails 7.2 but removed once again in Rails 8.0) and
commenting the `config.autoload_lib` option (we don't autoload the
`lib/` folder but use `app/lib/` instead).

[1] https://api.rubyonrails.org/v7.2/classes/Rails/Application.html#method-i-secret_key_base