Group Rails updates together in dependabot
Currently dependabot is failing to upgrade some of these gems. For
example, when there's a security issue in ActiveRecord or ActiveStorage,
dependabot fails because they need to be updated at once. We get
messages like:

```
Dependabot cannot update activestorage to a non-vulnerable version

The latest possible version that can be installed is 7.1.5.1 because of
the following conflicting dependencies:

rails (7.1.5.1) requires activestorage (= 7.1.5.1) via actionmailbox (7.1.5.1)
rails (7.1.5.1) requires activestorage (= 7.1.5.1) via actiontext (7.1.5.1)
rails (7.1.5.1) requires activestorage (= 7.1.5.1)

The earliest fixed version is 7.1.5.2.
```