Merge Sanitizable and Taggable concerns
The Sanitizable concern only makes sense for Taggable models since
commit 7bf4e4d61. Having two different concerns led to some
inconsitencies. The `Legislation::Process` class was including Taggable
but not Sanitizable (making it the only class where tags weren't
sanitized). On the othe hand, the `Budget::Phase` class still included
the `Sanitizable` concern even if it doesn't have tags, meaning
including it was useless.

We came up with this refactoring while trying to find places where we
could simplify code using the `normalizes` method included in Rails 7.1.
While doing so here turned out not to be trivial, we've decided to keep
this refactoring as part of this pull request.