Remove deprecated config option TODO: add commit reference when we merge the 7.2 branch
As mentioned in commit [TODO], now that we're using Rails 8.0, we're
removing this parameter. We were getting a warning:

```
DEPRECATION WARNING:
`config.active_job.enqueue_after_transaction_commit` is deprecated and
will be removed in Rails 8.1. This configuration can still be set on
individual jobs using `self.enqueue_after_transaction_commit=`, but due
the nature of this behavior, it is not recommended to be set globally.
(called from <top (required)> at app/mailers/application_mailer.rb:1)
```